pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "potholes"
version = "0.1.0"
description = "Pothole detection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "affine==2.4.0",
    "attrs==25.4.0",
    "branca==0.8.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "click-plugins==1.1.1.2",
    "cligj==0.7.2",
    "colorama==0.4.6",
    "contextily==1.7.0",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "folium==0.20.0",
    "fonttools==4.61.1",
    "geographiclib==2.1",
    "geopy==2.4.1",
    "humanfriendly==10.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "kiwisolver==1.4.9",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mercantile==1.2.1",
    "narwhals==2.15.0",
    "numpy==2.2.6",
    "optype==0.9.3",
    "packaging==25.0",
    "pandas==2.3.3",
    "pandas-stubs==2.3.3.251219",
    "pillow==12.1.0",
    "pip==25.3",
    "plotly==6.5.2",
    "pyparsing==3.3.1",
    "pyproj==3.6.1",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "rasterio==1.4.4",
    "requests==2.32.5",
    "scipy==1.15.3",
    "scipy-stubs==1.15.3.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "tqdm==4.67.1",
    "types-pytz==2025.2.0.20251108",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
    "urllib3==2.6.2",
    "wheel==0.45.1",
    "xyzservices==2025.11.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]